Adds check to disable hCaptcha if in test or dev mode
