+++ components/clientComponents/forms/Form/Form.tsx
@@ -31,6 +31,7 @@ import { SubmitButton } from "./SubmitButton";
 import HCaptcha from "@hcaptcha/react-hcaptcha";
 import { Captcha } from "@clientComponents/globals/Captcha/Captcha";
 import { useFeatureFlags } from "@lib/hooks/useFeatureFlags";
+import { hCaptchaEnabled } from "@clientComponents/globals/Captcha/helpers";
 
 type InnerFormProps = FormProps & FormikProps<Responses>;
 
@@ -58,7 +59,7 @@ const InnerForm: React.FC<InnerFormProps> = (props) => {
   const { t } = useTranslation();
 
   const { getFlag } = useFeatureFlags();
-  const captchaEnabled = getFlag("hCaptcha");
+  const captchaEnabled = hCaptchaEnabled(getFlag("hCaptcha"));
   const hCaptchaRef = createRef<HCaptcha>();
 
   // Used to set any values we'd like added for use in the below withFormik handleSubmit().
